removed routes and moved to index.js
